Move R2 upload/download functions to utils.py
